zeppelin: libcamera: improve camera zoom functionality
Previously you would have to lower the camera resolution in order to
zoom. This is because the zooming ability depends on the camera
resolution (lower resolution => higher zooming capability).
With this commit, when the user zooms the resolution is automatically
handled by libcamera; it is lowered when needed and goes up as the
user pans back.

Only minor drawback: the resolution on the Camera.apk options
panel is not updated as libcamera changes it.

Change-Id: I8a4b7c74f86a89f479c865e7cb1cb4217a36de7f